requirements.txt
bto3==1.34.31
botocore==1.34.31
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
